manifold/pyproject.toml
[project]
name = "manifold-mcp"
version = "0.1.0"
description = "Manifold Markets API as an MCP server -- powered by Invariant Protocol"
requires-python = ">=3.10"
dependencies = [
    "invariant-protocol @ git+https://github.com/jim-technologies/invariantprotocol.git#subdirectory=python",
    "pyyaml>=6.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gen"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
exclude = [
    "src/gen",
]
